Move authentication mechanism into the connector
When we first created the Redfish client, we decided to implement the
authentication process entirely in the root object. This way, we were
able to keep the connector and the resources loosely coupled.

Unfortunately, this design proved itself to be a bit too naive. Adding
support for automatic login after the active session expires would
mean that each resource would need to know how to perform the login
procedure.

To overcome the limitations of the existing system, we decided to move
the authentication parts to the connector, where all resources can
peruse it automatically.