Abstract away connector response
To give library users a more convenient and safer interface, we
abstracted the response into separate response class that has clearly
defined static members.

The main reason for adding this abstraction are semi-broken Redfish
implementations that sometimes return a non-JSON response. If the
end-user tries to access the JSON data in this situation, Redfish
client raises an exception, making our library highly inconvenient to
use.