Move cache into dedicated connector class
Previously, we had a per-resource cache, which was not optimal. For
example, we can access Redfish session via two different paths:
root.Links.Sessions and root.SessionService.Sessions. Even though both
calls access the same path, we hit the Redfish service twice, because
Links and SessionService caches are not shared.

Now that we placed the cache inside the connector, both calls share
the same cache, reducing the traffic in the Redfish service.