Add pylint configuration
In order to keep the code base relatively uniform, we introduced the
pylint configuration file that should catch most of the stylistic
offenses as well as some semantic errors. The format we are checking
is basically what PEP8 considers a good practice.

In order to start with a clean state, we also fixed all but one issue
pylint reported. Offense that we left actually marks a bug in our code
base and will be rectified once the bug is fixed.