Expose session authentication data
Since Redfish sessions are relatively long-lived, we can substantially
reduce the load on the Redfish API by reusing the same session and
associated authentication token. But in order to be able to reuse
authentication data, we need some way of accessing it.

This commit adds a property to the connector objects that contains
currently active set of authentication data. This data can be then
stored somewhere and reused when authenticating new connector.