Remove x-auth-token header before re-login
The connector part of the Redfish client library has the capability to
retry the request if the Redfish service that client is talking to
returns back status 401. Before the retry, connector tries to login in
order to refresh the authentication token.

It turns out that on older Dell systems (13G), sending an invalid or
expired token in x-auth-token header when creating new session also
returns 401, which causes the client to back out and abort the current
operation.

Updated code simply removes the x-auth-token header before trying to
create new session.