Lazy-load resources
Until now, we eagerly loaded Resource's content as soon as we
built it. This proved to be a bit awkward for resources with
many siblings, as accessing the contents of one such resource
resulted in fetching of the sibling resources' contents as well.

With this commit, lazy loading becomes the default behavior.
If this is not desired, the user may still use eager resources
by setting the `lazy_load` keyword argument of the client's
connect method to `False`.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>