Update CI/CD pipeline
Our pipeline was really rudimentary and only tested now-obsolete
Python version.

With changes in this commit, we bring the advertised Python support
and our testing setup back in sync.

Do note that the testing setup does not use pipenv anymore. The main
reason for this is multitude of issues in pipenv that failes to
properly detect Python version currently in use. We should probably
get rid of it from the rest of the project soonish since the upstream
support and development is flaky at best.

And yes, I am aware that I was the one who brought the pipenv into the
project. But I was young and naive back then. Now I (should) know
better ;)