Strip non-path components from location header when used for session id
Some RedFish implementations (HPE/iLO 4 and 5) were returning the full
URI in the "location" header. Logout would fail for these implementations
becuase the _session_logout() assumes just a path is returned. This is
an issue with the RedFish implementation but seems like a harmless fix
to include in the library.